ai-models/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# Для будущего развития ИИ:
# transformers==4.35.0
# torch==2.1.0
# pillow==10.0.1
# opencv-python==4.8.1
# numpy==1.24.3